Solvers

Flowpipe computation

CommonSolve.solve — Function

solve(prob::AbstractControlProblem, args...; kwargs...)

Solves the controlled problem defined by prob

Input

prob – controlled problem

Additional options are passed as arguments or keyword arguments; see the notes below for details. See the online documentation for examples.

Output

The solution of a reachability problem controlled by a periodic controller.

Notes

Use the tspan keyword argument to specify the time span; it can be:

a tuple,

an interval, or

a vector with two components.

Use the T keyword argument to specify the time horizon; the initial time is then assumed to be zero.

Use the alg_nn keyword argument to specify the solver for the controller.

While this function is written with a neural-network controlled systems in

mind, the type of the controller is arbitrary, as long as a function forward_network to analyze it is available.

Trajectory simulation

NeuralNetworkAnalysis.simulate — Function

simulate(cp::AbstractControlProblem, args...; kwargs...)

Simulate a controlled system for a family of random trajectories.

Input

cp – controlled problem

trajectories – (optional, default: 10) number of simulated trajectories

Output

An object of type EnsembleSimulationSolution.

Notes

This function uses the ensemble simulations feature from DifferentialEquations.jl.