Utility functions

Locating files

NeuralNetworkAnalysis.@modelpath — Macro

@modelpath(name)

Return the absolute path to file name relative to the executing script.

Input

model_path – folder name (ignored by default)

name – filename

Output

A string.

Notes

This macro is equivalent to joinpath(@__DIR__, name). The @modelpath macro is used in model scripts to load data files relative to the location of the model, without having to change the directory of the Julia session. For instance, suppose that the folder /home/projects/models contains the script my_model.jl, and suppose that the data file my_data.dat located in the same directory is required to be loaded by my_model.jl. Then,

# suppose the working directory is /home/julia/ and so we ran the script as
# julia -e "include("../projects/models/my_model.jl")"
# in the model file /home/projects/models/my_model.jl we write:
d = open(@modelpath("", "my_data.dat"))
# do stuff with d

In this example, the macro @modelpath("", "my_data.dat") evaluates to the string /home/projects/models/my_data.dat. If the script my_model.jl only had d = open("my_data.dat"), without @modelpath, this command would fail as julia would have looked for my_data.dat in the working directory, resulting in an error that the file /home/julia/my_data.dat is not found.

Reading networks

NeuralNetworkAnalysis.read_nnet_mat — Function

read_nnet_mat(file; key=nothing, act_key="activation_fcns")

Read a neural network stored in a .mat file and return the corresponding network in the format of NeuralVerification.jl.

Input

file – string indicating the location of the .mat file containing the neural network

key – (optional, default: nothing) key used to search the dictionary containing the controller; by default we search the top-level dictionary; a typical value is "controller"

act_key – (optional, default: "activation_fcns") key used to search the activation functions; typical values are "activation_fcns" or "act_fcns"

Output

A Network struct.

Notes

The following activation functions are supported:

RELU: "relu" (ReLU)

Identity: "linear" (Id)

NeuralNetworkAnalysis.read_nnet_yaml — Function

read_nnet_yaml(data::Dict)

Read a neural network from a file in YAML format (see YAML.jl) and convert it

Read a neural network stored in a .mat file and return the corresponding network in the format of NeuralVerification.jl.

Input

file – string indicating the location of the .mat file containing the neural network

key – (optional, default: "controller") key used to search the dictionary

Output

A Network struct.

Notes

The following activation functions are supported: identity, relu, sigmoid and tanh; see NeuralNetworkAnalysis.ACT_YAML.